requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==30.1.0
faker-vehicle==0.2.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
